+++ core/src/main/java/com/adobe/aem/commons/assetshare/content/renditions/AssetRendition.java
@@ -19,12 +19,12 @@
 
 package com.adobe.aem.commons.assetshare.content.renditions;
 
+import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.UnsupportedEncodingException;
 import java.net.URI;
-import java.net.URLEncoder;
+import java.net.URISyntaxException;
 import java.util.Optional;
 
 /**
@@ -39,21 +39,21 @@ public class AssetRendition {
     private String mimeType;
 
     public AssetRendition(URI uri, Long size, String mimeType) {
-        setBinaryUri(uri);
-        setSize(size);
-        setMimeType(mimeType);
+        this(uri.toString(), size, mimeType);
+
     }
 
     public AssetRendition(String uri, Long size, String mimeType) {
+        URI cleanURI = null;
+
         try {
-            uri = URLEncoder.encode(uri, "UTF-8");
-        } catch (UnsupportedEncodingException ex) {
-            log.warn("Unable to encode String URI [ {} ] using UTF-8. Continuing using URI unencoded...", uri);
+            cleanURI = cleanURI(uri.toString());
+        } catch (URISyntaxException e) {
+            log.warn("Unable to clean the URI [ {} ], using it as is.", uri, e);
+            cleanURI = URI.create(uri);
         }
 
-        uri = uri.replace("+", "%20");
-
-        setBinaryUri(URI.create(uri));
+        setBinaryUri(cleanURI);
         setSize(size);
         setMimeType(mimeType);
     }
@@ -88,6 +88,12 @@ public class AssetRendition {
         this.mimeType = mimeType;
     }
 
+    private URI cleanURI(String uri) throws URISyntaxException {
+        uri  = StringUtils.replace(uri, " ", "%20");
+        uri  = StringUtils.replace(uri, "/_jcr_content", "/jcr:content");
+
+        return new URI(uri);
+    }
 }